Check tests are still being run
